Vterm
really want to use the built-in term but I feel like the user-interface is
terrible. i could call term with my own function to set C-x as an escape prefix
just like in ansi-term, but when i press C-c in char-mode it activates the C-x
prefix map? vterm handles all this super intuitively and i can even call eval
and stuff.